# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/add": {
      "post": {
        "operationId": "AuthController_create",
        "summary": "create one new auth",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuthDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/page": {
      "get": {
        "operationId": "AuthController_findAll",
        "summary": "find auth by page",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isDel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/{id}": {
      "get": {
        "operationId": "AuthController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      },
      "patch": {
        "operationId": "AuthController_update",
        "summary": "update auth info",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/remove{id}": {
      "delete": {
        "operationId": "AuthController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/enable/{id}": {
      "patch": {
        "operationId": "AuthController_enable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/page": {
      "post": {
        "operationId": "PageController_create",
        "summary": "create one new page",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Page"
        ]
      }
    },
    "/page/page": {
      "get": {
        "operationId": "PageController_findAll",
        "summary": "find page list by page",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isDel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "hasAuth",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Page"
        ]
      }
    },
    "/page/{id}": {
      "get": {
        "operationId": "PageController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Page"
        ]
      },
      "patch": {
        "operationId": "PageController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Page"
        ]
      }
    },
    "/page/remove{id}": {
      "delete": {
        "operationId": "PageController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Page"
        ]
      }
    },
    "/page/enable/{id}": {
      "patch": {
        "operationId": "PageController_enable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Page"
        ]
      }
    },
...
    "/file/upload": {
      "post": {
        "operationId": "FileController_upload",
        "summary": "upload image file",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "File"
        ]
      }
    },
    "/file/upload/mass": {
      "post": {
        "operationId": "FileController_massUpload",
        "summary": "mass upload image files",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "File"
        ]
      }
    },
    "/file/page": {
      "get": {
        "operationId": "FileController_findAll",
        "summary": "find files by page",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isDel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "visibleRange",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "P",
                "O",
                "L"
              ],
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "avatar",
                "photo",
                "illustration"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "File"
        ]
      }
    },
    "/file/{id}": {
      "get": {
        "operationId": "FileController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "File"
        ]
      }
    },
    "/file/remove{id}": {
      "delete": {
        "operationId": "FileController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "File"
        ]
      }
    },
    "/file/enable/{id}": {
      "patch": {
        "operationId": "FileController_enable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "File"
        ]
      }
    },
    "/category": {
      "post": {
        "operationId": "CategoryController_create",
        "summary": "create one new category",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/category/page": {
      "get": {
        "operationId": "CategoryController_findAll",
        "summary": "find all category by page",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isDel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "hasPost",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/category/{id}": {
      "get": {
        "operationId": "CategoryController_findOne",
        "summary": "find one category by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isDel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "hasPost",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Category"
        ]
      },
      "patch": {
        "operationId": "CategoryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/category/remove/{id}": {
      "delete": {
        "operationId": "CategoryController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/category/enable/{id}": {
      "patch": {
        "operationId": "CategoryController_enable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
...
  },
  "info": {
    "title": "My API",
    "description": "The description of my API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "token": {
        "type": "apiKey",
        "scheme": "basic",
        "name": "access-token",
        "description": "user token authentication",
        "in": "headers"
      }
    },
    "schemas": {
...
      "CreateAuthDto": {
        "type": "object",
        "properties": {
          "authName": {
            "type": "string",
            "description": "auth name",
            "minLength": 2,
            "maxLength": 12
          },
          "authKey": {
            "type": "string",
            "description": "auth key",
            "minLength": 5,
            "maxLength": 24
          },
          "remark": {
            "type": "string",
            "description": "remark",
            "maxLength": 50
          },
          "path": {
            "type": "string",
            "description": "path"
          },
          "roles": {
            "description": "roles",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "authName",
          "authKey",
          "remark",
          "path",
          "roles"
        ]
      },
      "UpdateAuthDto": {
        "type": "object",
        "properties": {
          "authName": {
            "type": "string",
            "description": "auth name",
            "minLength": 2,
            "maxLength": 12
          },
          "authKey": {
            "type": "string",
            "description": "auth key",
            "minLength": 5,
            "maxLength": 24
          },
          "remark": {
            "type": "string",
            "description": "remark",
            "maxLength": 50
          },
          "path": {
            "type": "string",
            "description": "path"
          },
          "roles": {
            "description": "roles",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "authName",
          "authKey",
          "remark",
          "path",
          "roles"
        ]
      },
      "CreatePageDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "page title",
            "minLength": 2,
            "maxLength": 20
          },
          "path": {
            "type": "string",
            "description": "page path",
            "minLength": 4,
            "maxLength": 45
          },
          "sort": {
            "type": "number",
            "description": "page sort order",
            "minimum": 0,
            "maximum": 999
          },
          "component": {
            "type": "string",
            "description": "component path",
            "minLength": 4,
            "maxLength": 45
          },
          "isIframe": {
            "type": "string",
            "description": "is it iframe",
            "default": "1",
            "enum": [
              "0",
              "1"
            ]
          },
          "type": {
            "type": "string",
            "description": "page type",
            "default": "P",
            "enum": [
              "P",
              "B"
            ]
          },
          "isLink": {
            "type": "string",
            "description": "is it external link",
            "default": "1",
            "enum": [
              "0",
              "1"
            ]
          },
          "visible": {
            "type": "string",
            "description": "visible",
            "default": "0",
            "enum": [
              "0",
              "1"
            ]
          },
          "redirect": {
            "type": "string",
            "description": "redirect path",
            "minLength": 4,
            "maxLength": 45
          },
          "prefixIcon": {
            "type": "string",
            "description": "prefix icon",
            "minLength": 4,
            "maxLength": 20
          },
          "suffixIcon": {
            "type": "string",
            "description": "suffix icon",
            "minLength": 4,
            "maxLength": 20
          },
          "remark": {
            "type": "string",
            "description": "remark",
            "minLength": 4,
            "maxLength": 50
          },
          "auth": {
            "type": "string",
            "description": "auth id"
          }
        },
        "required": [
          "title",
          "path",
          "sort",
          "component",
          "isIframe",
          "type",
          "isLink",
          "visible",
          "redirect",
          "prefixIcon",
          "suffixIcon",
          "remark",
          "auth"
        ]
      },
      "UpdatePageDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "page title",
            "minLength": 2,
            "maxLength": 20
          },
          "path": {
            "type": "string",
            "description": "page path",
            "minLength": 4,
            "maxLength": 45
          },
          "sort": {
            "type": "number",
            "description": "page sort order",
            "minimum": 0,
            "maximum": 999
          },
          "component": {
            "type": "string",
            "description": "component path",
            "minLength": 4,
            "maxLength": 45
          },
          "isIframe": {
            "type": "string",
            "description": "is it iframe",
            "default": "1",
            "enum": [
              "0",
              "1"
            ]
          },
          "type": {
            "type": "string",
            "description": "page type",
            "default": "P",
            "enum": [
              "P",
              "B"
            ]
          },
          "isLink": {
            "type": "string",
            "description": "is it external link",
            "default": "1",
            "enum": [
              "0",
              "1"
            ]
          },
          "visible": {
            "type": "string",
            "description": "visible",
            "default": "0",
            "enum": [
              "0",
              "1"
            ]
          },
          "redirect": {
            "type": "string",
            "description": "redirect path",
            "minLength": 4,
            "maxLength": 45
          },
          "prefixIcon": {
            "type": "string",
            "description": "prefix icon",
            "minLength": 4,
            "maxLength": 20
          },
          "suffixIcon": {
            "type": "string",
            "description": "suffix icon",
            "minLength": 4,
            "maxLength": 20
          },
          "remark": {
            "type": "string",
            "description": "remark",
            "minLength": 4,
            "maxLength": 50
          },
          "auth": {
            "type": "string",
            "description": "auth id"
          }
        },
        "required": [
          "title",
          "path",
          "sort",
          "component",
          "isIframe",
          "type",
          "isLink",
          "visible",
          "redirect",
          "prefixIcon",
          "suffixIcon",
          "remark",
          "auth"
        ]
      },
...
      "CreateFileDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "file description",
            "maxLength": 101
          },
          "visibleRange": {
            "type": "string",
            "description": "file visible range",
            "enum": [
              "P",
              "O",
              "L"
            ],
            "default": "P"
          },
          "type": {
            "type": "string",
            "description": "file type",
            "enum": [
              "avatar",
              "photo",
              "illustration"
            ],
            "default": "illustration"
          }
        },
        "required": [
          "description",
          "visibleRange",
          "type"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "categoryName": {
            "type": "string",
            "description": "标签名称"
          },
          "remark": {
            "type": "string",
            "description": "标签备注"
          },
          "iconName": {
            "type": "string",
            "description": "小图标"
          },
          "postIds": {
            "description": "Post ids",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "categoryName",
          "remark",
          "iconName"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}